requirements.txt
ccxt==2.9.11
pandas==1.5.3
numpy==1.24.3
matplotlib==3.7.1
TA-Lib==0.4.25
scikit-learn==1.2.2
tensorflow==2.12.0
Flask==2.3.2
requests==2.30.0
python-dotenv==1.0.0
plotly==5.14.1